Fix reverse transactions tests
